add confirmation dialog when delete category

+++ apps/admin/src/app/categories/categories-list/categories-list.component.ts
@@ -1,6 +1,6 @@
 import { Component, OnInit } from '@angular/core';
 import { CategoriesService, Category } from '@bluebits/products';
-import { MessageService } from 'primeng/api';
+import { MessageService, ConfirmationService } from 'primeng/api';
 
 @Component({
   selector: 'admin-categories-list',
@@ -15,6 +15,7 @@ export class CategoriesListComponent implements OnInit {
   constructor(
     private categoriesService: CategoriesService,
     private messageService: MessageService,  
+    private confirmationService: ConfirmationService
   ) { }
 
   ngOnInit(): void {
@@ -26,14 +27,24 @@ export class CategoriesListComponent implements OnInit {
   }
 
   deleteCategory(category: Category) {
-    this.categoriesService.deleteCategory(category.id).subscribe({
-      next: () => {
-        this.messageService.add({severity: 'success', summary: 'Success', detail: `Category ${category.name} is deleted`});
-        this.fetchCategories();
-      },
-      error: () => this.messageService.add({severity: 'success', summary: 'Success', detail: 'Category is not deleted'})
+    this.confirmationService.confirm({
+      message: `Do you want to delete this '${category.name}' category?`,
+      header: `Delete '${category.name}' category`,
+      icon: 'pi pi-exclamation-triangle',
+      accept: () => {
+        this.categoriesService.deleteCategory(category.id).subscribe({
+          next: () => {
+            this.messageService.add({severity: 'success', summary: 'Success', detail: `Category ${category.name} is deleted`});
+            this.fetchCategories();
+          },
+          error: () => this.messageService.add({severity: 'success', summary: 'Success', detail: 'Category is not deleted'})
+        });
+      },      
     });
   }
+
+
+  
 
   editCategory() {
